Recognizing road lanes
This project features a lane recognition system, devised for a camera
placed in a car.It is capable of differentiating between lanes that are
pointed towards our current direction and between lanes in the opposite
direction. Currently it processes images converted from the actual video
made by camera placed behind the front window in a car.